+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+- ContentTest - Check length of array or string
+- ContentTest - Check minimum length of array or string of response body
+- ContentTest - Check maximum length of array or string of response body
+- ContentTest - Check minimum length of array or string of response header
+- ContentTest - Check maximum length of array or string of response header
+
 ## v1.12.2 - (2021-12-15)
 
 - Added better handling of Postman API errors